middlewares added

+++ controllers/cards.js
@@ -1,68 +1,66 @@
 const mongoose = require('mongoose');
 const Card = require('../models/card');
-const {
-  VALIDATION_ERROR,
-  NOT_FOUND_ERROR,
-  REFERENCE_ERROR,
-} = require('../errors/errorsCodes');
+const NotFound = require('../errors/notFound');
+const Forbidden = require('../errors/forbidden');
+const Validation = require('../errors/validation');
 
-module.exports.getCards = (req, res) => {
+module.exports.getCards = (req, res, next) => {
   Card.find({})
     .populate('owner')
     .then((cards) => res.send({ data: cards }))
-    .catch(() => res.status(REFERENCE_ERROR).send({ message: 'Произошла ошибка по умолчанию' }));
+    .catch(next);
 };
 
-module.exports.createCard = (req, res) => {
+module.exports.createCard = (req, res, next) => {
   const { name, link } = req.body;
-
   Card.create({ name, link, owner: req.user._id })
     .then((card) => res.send({ data: card }))
     .catch((err) => {
-      if (err instanceof mongoose.Error.ValidationError) {
-        res.status(VALIDATION_ERROR).send({ message: 'Переданы некорректные данные при создании карточки' });
+      if (err instanceof mongoose.Error.Validation) {
+        next(new Validation('Переданы некорректные данные при создании карточки'));
       } else {
-        res.status(REFERENCE_ERROR).send({ message: 'Произошла ошибка по умолчанию' });
+        next(err);
       }
     });
 };
 
 module.exports.deleteCardById = (req, res, next) => {
-  Card.findByIdAndRemove(req.params.cardId)
+  Card.findById(req.params.cardId)
     .then((card) => {
       if (!card) {
-        res.status(NOT_FOUND_ERROR).send({ message: 'Карточка с указанным _id не найдена.' });
+        throw new NotFound('Карточка с указанным _id не найдена');
       } else {
-        next(res.send({ data: card }));
+        const cardOwner = card.owner.toString();
+        if (req.user._id !== cardOwner) {
+          throw new Forbidden('У вас нет прав на удаление данной карточки');
+        } else {
+          return Card.findByIdAndRemove(req.params.cardId);
+        }
       }
     })
-    .catch((err) => {
-      if (err instanceof mongoose.Error.CastError) {
-        res.status(VALIDATION_ERROR).send({ message: 'Переданы некорректные данные для удаления карточки' });
-      } else {
-        res.status(REFERENCE_ERROR).send({ message: 'Произошла ошибка по умолчанию' });
-      }
-    });
+    .then((card) => {
+      res.send({ data: card });
+    })
+    .catch(next);
 };
 
 module.exports.likeCard = (req, res, next) => Card.findByIdAndUpdate(
   req.params.cardId,
   { $addToSet: { likes: req.user._id } },
   { new: true },
-  { runValidators: true },
 )
   .then((card) => {
     if (!card) {
-      res.status(NOT_FOUND_ERROR).send({ message: 'Переданы некорректные данные для постановки лайка.' });
+      throw new NotFound('Передан несуществующий _id карточки');
     } else {
       next(res.send({ data: card }));
     }
   })
   .catch((err) => {
     if (err instanceof mongoose.Error.CastError) {
-      res.status(VALIDATION_ERROR).send({ message: 'Передан несуществующий _id карточки.' });
+      next(new Validation('Переданы некорректные данные для постановки лайка'));
     } else {
-      res.status(REFERENCE_ERROR).send({ message: 'Произошла ошибка по умолчанию' });
+      next(err);
     }
   });
 
@@ -73,15 +71,15 @@ module.exports.dislikeCard = (req, res, next) => Card.findByIdAndUpdate(
 )
   .then((card) => {
     if (!card) {
-      res.status(NOT_FOUND_ERROR).send({ message: 'Передан несуществующий _id карточки.' });
+      throw new NotFound('Передан несуществующий _id карточки');
     } else {
       next(res.send({ data: card }));
     }
   })
   .catch((err) => {
     if (err instanceof mongoose.Error.CastError) {
-      res.status(VALIDATION_ERROR).send({ message: 'Переданы некорректные данные для снятия лайка.' });
+      next(new Validation('Переданы некорректные данные для снятия лайка'));
     } else {
-      res.status(REFERENCE_ERROR).send({ message: 'Произошла ошибка по умолчанию' });
+      next(err);
     }
   });
